Rewriting base Code

+++ Assets/Scripts/GameManager.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using System.Diagnostics;
 using UnityEngine;
 
@@ -18,6 +15,6 @@ public class GameManager : MonoBehaviour
 
     private void OnApplicationQuit()
     {
-        process.Kill();
+        if(!Application.isEditor) process.Kill();
     }
 }
